Move /modroles into own folder and refactor.
